Refactor rule system loading, start with metadata
It started to make a lot of sense that we didn't really want to specify
what files should be named to load rule systems. Instead, we should load
all the .json files for a system and then parse the RuleSystem struct
from that. This commit enables that and also begins loading the rule
sytem metadata.