Upgrade to optimus 0.5
We were on optimus `0.3`, and the latest version is `0.5`. Nothing
seems to have broken with the upgrade. It does seem that optimus may
be at this point unmaintained. It's last update was over a year ago.
However, I might not be one to speak given I took a 2 year gap on this
project.