Ensure CLI results for rolling dice are printed as integers
`IO.inspect` by default treats lists of integers as `charlists`. This
means certain integer combinations result in the list being printed as
`charlist`. This was rather confusing, as getting `~c"\a\n"` when
rolling `2d10` doesn't make much sense (`~c"\a\n"` == `[7, 10]`). To
force `IO.inspect` to read the list as a list of integers, we had to
pass `charlists: :as_lists` in the opts of the function.