Add `faker` as a dependency "temporarily"
We the ability to generate names (for characters). Ideally we'd have a
name generator that is fanatasy geared and/or tune-able by setting. It
looks like, if we want that, we'll have to build it. However, that is
out of scope for this branch of work. I plan to revisit it later/soon
though as I find that a rather interesting problem.