Flatten `handle_show_subcommands` into a patternmatched parent
With the removal of `gen` CLI subcommand of `systems` in the previous
commit, the only sub command being handled by the "general"
`handle_systems_subcommands` method was `:show`. As such, it made sense
to flatten the child function being called into a patterned matched on
`:show` version of `handle_systems_subcommands`.

Of note, this seems to be a generally useful pattern, that is having
handling functions pattern match on the command. If we expanded this
paradigm to subcommands, we'd approach a system which each
command + sub-command combo chain has it's own individual handler. If
that becomes the rule, it might make sense to extend `optimus` such that
leaf commands simply define their handler in the spec. This would
elimate the need for functions which mostly seem to becoming routing.