Loosen typing on `character_exists?` to allow for slugs as parameters
This was done because previously `character_exists?` could only used by passing in a `Character` struct. This has it's uses, but unfortunately meant that if you only had a character slug, you couldn't check whether or not a character exists. As we're trying to implement showing a character via the CLI, it's nice to check if the provided slug exists before trying to load it.