Add CLI command `characters show` to show a saved character
Of note, we had to duplicate some helper functions from
`characters_test.exs` to `custom_parsers_test.exs`. It'd be desirable if
these could be shared utility methods that multiple tests can use.
However to do so we either need to include the test helper functions in
the package libary code _or_ create some test macros. The former is
undersireable as I don't believe testing helpers should generally be
distributed as part of the library. The latter is more technical than we
need at this point in time. Sometime in the future we are likely to move
to the latter option.